Add files via upload
## 1. Neural Network for Digits Classification
This project focuses on creating a feedforward neural network using Python and NumPy. It starts with implementing a simple perceptron and gradually builds up to more complex concepts. The project includes:

- Training a neural network on a toy dataset for digits classification.
- Implementing activation functions, including the sigmoid function.
- Weight initialization and forward propagation.
- Backpropagation to update weights.
- Training iterations with mean squared error analysis.
- Demonstrating batch training for improved efficiency.

To run this code, you'll need Python and NumPy installed. The provided Jupyter Notebook guides you through the entire process.

## 2. Two-Layer Neural Network
This part introduces a more complex problem that cannot be solved with a single-layer perceptron. It demonstrates the need for a two-layer neural network. Key features include:

- Defining a new training set for XOR and NOT logic operations.
- Weight initialization for the two-layer network.
- Implementing forward and backward propagation.
- Training the network to solve the XOR and NOT operations.


## 3. Handwritten Digits Classification
In this part, I use a neural network to classify handwritten digits using a softmax activation function and cross-entropy loss. The key components include:

- Loading and visualizing the "load_digits" dataset.
- Implementing the softmax activation function.
- Using cross-entropy loss for training.
- Weight initialization and forward propagation.
- Training the model on the digit dataset to achieve high accuracy.

Feel free to reach out if you have any questions or feedback on these projects. I'm constantly working on new projects and improving my skills, so stay tuned for updates to this portfolio!

Thank you for visiting my GitHub portfolio.